Move Quartz extensions into Bg.Jobs project
