+++ OutOfSchool/OutOfSchool.BackgroundJobs/Extensions/Startup/AverageRatingExtensions.cs
@@ -1,10 +1,11 @@
-﻿using OutOfSchool.BackgroundJobs.Config;
+﻿using Microsoft.Extensions.DependencyInjection;
+using OutOfSchool.BackgroundJobs.Config;
 using OutOfSchool.BackgroundJobs.Jobs;
 using OutOfSchool.BusinessLogic.Services.AverageRatings;
 using OutOfSchool.Common.QuartzConstants;
 using Quartz;
 
-namespace OutOfSchool.WebApi.Extensions.Startup;
+namespace OutOfSchool.BackgroundJobs.Extensions.Startup;
 
 public static class AverageRatingExtensions
 {
