+++ OutOfSchool/OutOfSchool.BackgroundJobs/Extensions/Startup/NotificationExtensions.cs
@@ -1,11 +1,12 @@
-﻿using OutOfSchool.BackgroundJobs.Config;
+﻿using Microsoft.Extensions.DependencyInjection;
+using OutOfSchool.BackgroundJobs.Config;
 using OutOfSchool.BackgroundJobs.Jobs;
 using OutOfSchool.BusinessLogic.Services.LicenseApprovalNotification;
 using OutOfSchool.BusinessLogic.Services.NotificationsClearing;
 using OutOfSchool.Common.QuartzConstants;
 using Quartz;
 
-namespace OutOfSchool.WebApi.Extensions.Startup;
+namespace OutOfSchool.BackgroundJobs.Extensions.Startup;
 
 public static class NotificationExtensions
 {
