+++ OutOfSchool/OutOfSchool.BackgroundJobs/Extensions/Startup/StatisticReportExtensions.cs
@@ -1,10 +1,11 @@
-﻿using OutOfSchool.BackgroundJobs.Config;
+﻿using Microsoft.Extensions.DependencyInjection;
+using OutOfSchool.BackgroundJobs.Config;
 using OutOfSchool.BackgroundJobs.Jobs;
 using OutOfSchool.BusinessLogic.Services.StatisticReports;
 using OutOfSchool.Common.QuartzConstants;
 using Quartz;
 
-namespace OutOfSchool.WebApi.Extensions.Startup;
+namespace OutOfSchool.BackgroundJobs.Extensions.Startup;
 
 public static class StatisticReportExtensions
 {
